requirements.txt
Flask
Pykka
Hashids
Rx
RxPY
kafka-python